reworked init_db.sql v 1.1.1

+++ src/main/resources/init_db.sql
@@ -1,6 +1,6 @@
 CREATE TABLE books (
   id BIGINT NOT NULL AUTO_INCREMENT,
   title VARCHAR(255),
-  price INT,
+  price BIGINT,
   PRIMARY KEY (`id`)
 );
